TMC2130: handle DIAG correctly when entering STST
TMC2130 clears DIAG when the overload condition is resolved, not when
moving off an endstop. This means that when the motor is not moving DIAG
should be cleared.

Transition to standstill implies this, so handle this situation by
adding a couple of helper functions to set/clear DIAG consistently
and use them when StandStill is detected.

Add tests to check this condition.

Also cancel the StandStill timer when the object is destroyed to avoid
callbacks on uninitialized instances, causing crashes during quit.